Rename prmopt and modal
