+++ packages/web/src/pages/chat-page/components/CreateChatPrompt.module.css
